requirements.txt
# Web framework
Flask

numpy

piper-tts
# Additional dependencies
python-dotenv

Werkzeug

click
